refactor(#3038): split health.rs into directory modules (verbatim moves)
health.rs becomes the module root (HealthRegistry core + re-export
surface, 402 prod LoC). The 2.9k-line monolith body moves to six flat
health/ submodules as verbatim function moves:

  * runtime_resolve — resolve_bot_http/fetch_channel_name + the
    direct-meeting runtime/shared resolver pair (kept as two copies;
    dedupe deferred to Phase B)
  * headless_turn — headless agent-turn reserve/start API +
    start_direct_meeting
  * send_target — /api/discord/send target parsing/resolution
  * send_gate — send authorization ladder (#2047 Findings 7/9/10) +
    caller-class label gate
  * manual_delivery — manual outbound delivery + #2363 dedupe protocol
  * send_api — handle_send / handle_senddm body parsing

Body edits are limited to the four sanctioned mechanical categories:
(1) path re-spelling (super::router/meeting/settings/agentdesk_config/
provider_handles_channel/resolve_discord_token_by_hash now spelled from
the submodule depth), (2) private->pub(super) promotions for the
cross-submodule seams (visible set unchanged: health + descendants),
(3) //! extraction headers + replicated imports, (4) one test-rewire
use line in manual_v3_delivery_tests (HealthRegistry/handle_send via
the root re-export path). HealthRegistry field visibility is untouched
(descendant privacy, snapshot.rs/recovery.rs precedent); external
callers are unchanged — the root re-exports pin every existing
health::* path at its original visibility.

Test set identity: 63 tests under services::discord::health before and
after the split.
